backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-dotenv==1.2.1
pydantic==2.12.5
sendgrid==6.12.5
scikit-learn==1.8.0
nltk==3.9.2
numpy==2.3.3
openai==2.16.0
google-generativeai==0.8.6
